third_party/blink/web_tests/flag-specific/disable-layout-ng/external/wpt/css/CSS2/linebox/inline-negative-margin-001-expected.txt

This is a testharness.js-based test.
PASS [data-expected-height] 1
FAIL [data-expected-height] 2 assert_equals: 
<inline-block data-expected-height="10">123 <span style="margin-left: -8ch">123 </span></inline-block>
height expected 10 but got 20
PASS [data-expected-height] 3
PASS [data-expected-height] 4
PASS [data-expected-height] 5
FAIL [data-expected-height] 6 assert_equals: 
<div class="w4" data-expected-height="10"><img class="w5"><span style="margin-left: -5ch">123</span></div>
height expected 10 but got 20
PASS [data-expected-height] 7
PASS [data-expected-height] 8
PASS [data-expected-height] 9
FAIL [data-expected-height] 10 assert_equals: 
<div class="w4" data-expected-height="10"><img class="w5"><img class="w1" style="margin-left: -2ch"></div>
height expected 10 but got 20
PASS [data-expected-height] 11
PASS [data-expected-height] 12
PASS [data-expected-height] 13
Harness: the test ran to completion.
